Rename HostFxrProbe to StubTaskHarness
The original name conflated two ideas: `probe` suggested the project
was discovering something about hostfxr (it isn't), and the project
itself isn't really a probe but the smallest possible stand-in for an
MSBuild host. Rename and reword the doc comments accordingly:

* Folder/csproj: tests/HostFxrProbe -> tests/StubTaskHarness.
* Class/namespace: HostFxrProbe -> StubTaskHarness.
* MSBuild target name: _CollectHostFxrProbeOutputs -> _CollectStubTaskHarnessOutputs.
* Output dir in test bin: probes/ -> harness/.
* Source comments updated to `harness` where appropriate; the verb
  `probe` is kept where it correctly describes loader behavior.

No behavior change.
